Extract TokenInterceptor
